py-service/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
pdf2image
pytesseract
spacy
scikit-learn
joblib
datefinder 